32 bit conversion

+++ shared_common.h
@@ -35,6 +35,6 @@ using f64 = double;
 
 inline u32 get_tick_count() {
     using namespace std::literals;
-    return std::chrono::system_clock::now().time_since_epoch() / 1s;
+    return static_cast<u32>(std::chrono::system_clock::now().time_since_epoch() / 1s);
 }
